// src/api/generated/model/apiResponseCandidateDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { CandidateDto } from "./candidateDto";

export interface ApiResponseCandidateDto {
  success?: boolean;
  message?: string;
  data?: CandidateDto;
  errors?: string[];
  timestamp?: string;
  path?: string;
}