// src/api/generated/model/apiResponseExamSessionListResponse.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ExamSessionListResponse } from "./examSessionListResponse";

export interface ApiResponseExamSessionListResponse {
  success?: boolean;
  message?: string;
  data?: ExamSessionListResponse;
  errors?: string[];
  timestamp?: string;
  path?: string;
}